MYR-1921: added new endpoint to get wallet address
