+++ src/controllers/index.ts
@@ -11,7 +11,6 @@ export * from './network-currency.controller';
 export * from './notification.controller';
 export * from './people.controller';
 export * from './post.controller';
-export * from './post-walletaddress.controller';
 export * from './report.controller';
 export * from './report-user.controller';
 export * from './storage.controller';
@@ -31,4 +30,5 @@ export * from './user-social-media.controller';
 export * from './user-wallet.controller';
 export * from './vote.controller';
 export * from './wallet.controller';
+export * from './walletaddress.controller';
 export * from './server.controller';
